refactor: extract dashboard metric sections
